requirements.txt
python-mysql
sqlalchemy
pycrypto
redis
